Initial conversion into python 3

+++ neuralModel.py
@@ -23,13 +23,13 @@ def run():
             sample_rate=Fs, frame_size=frame_size)
 
         # Inner hair cell activity
-        model.ihc = nengo.Node(output=model.filterbank, size_out=freqs)
+        model.ihc = nengo.Node(output=model.filterbank, size_out=int(freqs))
 
         # Cochlear neurons projecting down auditory nerve
-        model.auditory = nengo.Ensemble(n_neurons=freqs * neurons_per_freq, dimensions=freqs)
+        model.auditory = nengo.Ensemble(n_neurons=int(freqs * neurons_per_freq), dimensions=int(freqs))
         nengo.Connection(model.ihc, model.auditory)
 
-        model.command = nengo.Node(asr_input.step, size_in=freqs, size_out=1)
+        model.command = nengo.Node(asr_input.step, size_in=int(freqs), size_out=1)
         nengo.Connection(model.auditory, model.command)
